+++ src/app/use-cases/__test__/send-signature.spec.ts
@@ -3,7 +3,7 @@ import * as E from "fp-ts/Either";
 import * as TE from "fp-ts/TaskEither";
 import { validate } from "@pagopa/handler-kit/lib/validation";
 import { InvalidEntityError } from "../../../error/invalid-entity";
-import { SendSignatureRequestBody } from "../../../infra/azure/functions/send-signature-request";
+import { SendSignatureRequestBody } from "../../../infra/azure/functions/send-signature-request-message";
 import { signatureRequestNotFoundError } from "../../../signature-request/signature-request";
 import { mockGetProduct, mockGetSignatureRequest } from "./mock";
